Added Front matter
On all the MD files, we have added code on top of the page (this is called front matter). We need this to render the page on the website.

On the front matter you will see Images: In there add the images you want to appear on the website project page.